Auth:  add ToBase32 and FromBase32 functions to EncryptionKey
If Stargate is to allow anonymous access, we need to be able to
pass encyrptionkeys in via URL.  The easiest way to do that is to make
it part of the domain name, so it has to be alphanumeric.

It was decided that RFC 4648 was the best approximation of
alphanumeric to use.  This PR allows the conversion of EncryptionKeys
to and from this Base32 encoding.

It was also decided to reduce the size of the key from 32 bytes to 16.
The main reason was to reduce the encoded key size.

While examining byte array sizes I realized that we reused nonces.
I altered the code to make this reuse more obvious.

Finally, per Jeff's request, I've also implemented a version byte
in the non-encoded EncryptionKey as described in the design doc.
base64(version byte || encryption key)
=

Change-Id: I40433d736c6133bdb4ea2a9e87e6cbb639944a6b