Rename /scripts to /jenkins
While reviewing https://review.dev.storj.io/c/storj/gateway-mt/+/3296
it occured to me that the only script there would create a lot of mess
if it were run outside Jenkins.  This rename seems like it helps
clarify the purpose of the script so that someone doesn't have to dig
into the code to understand its use.

Change-Id: I211283b2f931cad820a0ff7b5258f76b7bca97fc