Jenkins: Run awscli tests
This removes the awscli testing that we did during the `test-mint.sh`
setup, and moves it to a `test-aws.sh` file. This is run in its own
docker container so that we can easily run any number of integration
tests using the awscli.

As a first iteration, we run several basic commands using both
path-style and virtual-host style requests.

Change-Id: I50b6cfb7388ccf0ed9fdc49cbac92e7fdf3792a9