gateway-mt:  use tagsql
Tagsql sends monitoring information about our SQL usage via Monkit.
We need this for monitoring, and as a prerequisite for migrations.

The problem with this is that tagsql is from the storj.io/storj
repo.  We probably don't want to include all of storj.io here, so
until we move tagsql and dbutil out of storj.io/storj, it was
recommended that we duplicate them here with a README explanation.

The perl code in gen.sh is copied from satellitedb, so it's mostly
not introducting new things.

Change-Id: I614de4c73607ecbd10a96475e2f33fa0b5abf0b1