gateway-mt: add TLS to new S3 compatible server
We need TLS, including support for both a base domain and a wildcard
domain AKA virtual-path style.  This change allows any number of
certificates to be loaded from a directory.

One of the harder things to nail down about this PR was how to handle
binding parameters to structs, as much of the new server data is
duplicated from Minio.  I decided to put it all in the root, except
if it already existed, so as to not change existing logic.

Change-Id: Ia8c24ea342c3356a29cdb52fbdad48bb0f5a2904