gateway-mt: use Minio router as fallback
Remove the newRun CLI options because it'd be too hard to test.
Instead we take a progressive approach, to allow us to replace
routes one at a time.

Also, exploit some changes to Minio we made so that we can
create an IAM Store directly, rather than relying on environment
variables to convey settings.

Also, I cleaned up some inconsistencies in the README and fixed
some spelling errors.

Change-Id: I234e141573ed34309e4a85b54d323b63ba1b8422