gateway: Skip extra StatBucket on non-error path
Because many of our metainfo endpoints do not return a proper error when
the bucket is missing, we have previously had a manual check to stat the
bucket at the beginning of many operations.

With this change, in many methods, we now do not perform this check
unless we have already encountered an error.

Change-Id: Ibb8bf84e218c5a21132aaf092618dcd4533c2d58