gateway-mt:  Add check-downgrades to Jenkins
Check-downgrades had been run in the CI repo - against all the other
repos.  This caused an issue when Egon bumped the CI repo, because of an
apparent downgrade (not an actual one, a change from a tag to master).

Its best to check for downgrades in Gateway-MT directly as well.

Change-Id: Ic548c33f5fa13f51416a1db1938ff0080f6fa4d1