gateway-mt/auth: use satellite ids instead of addresses for inclusion list
We are changing the domain names of our satellites. For instance,
us-central-1.tardigrade.io is becoming us1.storj.io.

The identity of the satellite is best represented by its node id,
which cannot change. Because the satellite address can change, it
is not how we should identify satellites.

We need to roll this out so we can change the domain names of our
satellites and have the approval list work for access grants with
both the old and new satellite dns names.

All of our auth services already use node ids in their approval
list, so this change is both necessary and sufficient to fix the
problem.

For the gateway, all satellites support multipart now, so instead
of fixing that code, we can just take it out.

Change-Id: I59fc858ef87be26a12de23104b7b6c76ad829462