gateway-mt:  Merge experimental code
Caleb had an 'experimental' directory of code in Gateway-MT, and with
new employees starting, it seems prudent to merge that code into the
active codebase to not confuse people.

Overall, Caleb's additions used a lot of extra structs that don't offer
much benefit, and they stored a lot of things in the context.  I've
attempted to merge these things into the Server struct, so most of the
changes appear there.

In addition, we've begun to realize that its difficult to track what
endpoints Gateway-MT supports.  I begun to break these endpoints into
their own files for clarity and shorter files.  This leaves the main
server.go file with helper functions that more clearly show what it
does, while pushing the endpoint implementations into their own files.
The remaining functions to be ported away from Minio now have a
unimplemented.go file to see clearly what remains un-ported.

Unfortuntely, these functions are not yet operable on their own.  We
need to be able to get authorization working correctly first, as the
Server.GetUplinkProject we extracted from 'experimental' tries to find
the auth credentials in the context, but currently nothing saves these
credentials to the context.  That will come in another PR.

Error handling and mapping has been the real problem exposed by the
Gateway-MT logs.  I've created an errors.go file where all server
errors can be mapped to S3 errors.

Finally, I've begun working toward unwinding the extent to which we've
forked Minio.  The hard fork of Minio wasn't a problem so much as the
very large sections of code we removed from it.  We've learned over
time that a lot of 3rd party tools called some of the endpoints we
removed in the Minio hard fork.  In addition, a large amount of our
additions to minio code were simply for exposing unexported methods.
I've added an expose.go file which uses the linkname feature / trick
to import methods from Minio without altering the Minio codebase.
This might allow us to maintain a soft fork of Minio instead.

Change-Id: I7d4d294fe085680baef0c44c50526e408deb44fc