auth: improve tests
Tests in `satellitelist_test.go` have been written in a way that might
result in missing some edge cases. Specifically, there was no check for
nil error in test cases that didn't expect an error to be returned.

This change rewrites tests in `satellitelist_test.go` to account for
that and makes `satellitelist_test.go` and `auth/memauth` tests more
readable.

Change-Id: If76ca31b65d6c4a8074a615ecd8975d6af12e124