gateway-mt: expose Minio internals using go:linkname workaround
Traditionally, Gateway-MT relied on changing Minio's internals to
be able to access its router and authorization logic.  As much as
possible, we'd like to move that into Gateway-MT (at least for a while),
so that it makes it easier to for Gateway-MT and Gateway-ST to both use
the same version of Minio.

This change exposes or duplicates some of these Minio internals.

Change-Id: If7f4f24df9158ca08a8fc47362c750258952bb04