gateway-mt:  clarify secret-key data flow in signature.go
We copied signature.go over from Minio, but it appears that there were
some parts that weren't right from previous authors.  We must use the
Auth Service to get the secret key in order to validate the signature of
an S3 request.  Although we don't have tests for this yet, this change
at least makes the data flow clear.

Change-Id: Ic4ad67508db91cb45f71677b3ea808cec8a5a414