gateway-mt: ensure jenkins make binaries target correctly fails
make for loops continue regardless of the exit code of commands run
in the loop, with the last loop determining the actual
exit code of the make command. so for example: the first loop fails,
then the last one succeeds, this means the target is considered a
success.

exiting early in the loop will ensure we correctly fail.

Change-Id: Ie64211772c9fcc12c1f1fa44110d7137957badcb