auth/sqlauth: Use AS OF SYSTEM TIME when Cockroach
We are having quite a few outages with the production Auth services and
we have diagnosed that are attributed to CRDB queries that get aborted
due to transaction contentions.

Using 'AS OF SYSTEM TIME' CRDB clause will allow not to contribute to
the transaction contention when retrieving the secret key and access
grant which is the most used operation by the Auth service.

This commit has changed the key/value store backed by SQL to distinguish
between Postgres and Cockroach for only using the CRDB features when the
used DB is CRDB and it has also added a test to verify that works for
both because the changes have changed a DBX function call by a raw
query.