auth/sqlauth: Use AS OF SYSTEM TIME when Cockroach
We are having quite a few outages with the production Auth services and
we have diagnosed that are attributed to CRDB queries that get aborted
due to transaction contentions.

Using 'AS OF SYSTEM TIME' CRDB clause will allow not to contribute to
the transaction contention when retrieving the secret key and access
grant which is the most used operation by the Auth service and falling
back to execute the query without it when it returns no records.

This commit has also added some tests to verify that works for Postgres
and CRDB and make ensure that the CRDB query which uses 'AS OF SYSTEM
TIME' clause is called.

Change-Id: Icc7c39b1f2b462d4a2dcdd0ebf0f5c68134ce5a4