pkg/server:  make CopyObject return NotImplemented configurable
Gateway-ST implements CopyObject, but Gateway-MT has this code commented
out because if we allow it, we'd probably get hit with really expensive
global rename / move operations, so we want this to be done on the
satellite side instead.

This moves it to behind a config flag, so that they can share the same
logic.  Also, this allows us to fix the Mint tests so we run all the
copy object tests again.

Note that there's a time-based fudge-factor in these tests, and it
appears to need more time on CockroachDB.

Change-Id: I3998d6d7a08117eb6fddf0f29ce8c65c40cafbcf