testsuite/server: fix data race
`context.Context` passed down along with object layer is no longer
overwritten to avoid data race that arises when `testplanet` starts all
the nodes that use the context we overwrite in the meantime. We don't
need to overwrite the context and only build on it. This change
implements this behaviour in all places in `testsuite`.

Change-Id: I40e3ced30733b85537651c252d33910c03bc1529