pkg/server: Fix splitting port from IP v6
The function that get the IP from the HTTP request didn't work properly
if the remote address was an IP v6 when trying to remove the port part.

This commit has addressed the bug and added several IP v6 tests cases.

The function has been split to have the logic of extracting the IP from
headers into a separated one because it will be useful for just only
use that part for the new IP rate limiting that we plan to add.

It has also added a comment in the command-line help of the Gateway-MT
for the trusted IP list that they have to be provided without the port
part.

Change-Id: Ib34efe2ece4af6852540137fca962a92326978c5