{pkg/server, testsuite/server}: fix empty bucket deletion
This change fixes empty bucket deletion with pending multipart uploads.
Other S3 implementations (like AWS S3 or MinIO) allow deletion of such
buckets, but with libuplink, we need to explicitly force bucket
deletion. Previously we errored out that the bucket was not empty
because we didn't check for that condition and haven't issued forced
deletion in that situation.

Change-Id: I75c03a3b02ffa8b920dd8198785e503b5a74ba23