pkg/server: fix convertError to return correct SlowDown error
Minio internally call toAPIError(), and minio.SlowDown already is an API
error.  Therefore we don't want to return SlowDown, we want to return
something that maps to SlowDown.  Our options are all terribly named, so
we include a comment to help explain.

This should allow Minio to return 503s where it previously returned
500s.

Change-Id: I43b5bad7daa61dfbbdf6930feb1797c0bcc312df