pkg/linksharing:  add permissive CORS handling
A user request came in for supporting CORS in linksharing, which can be
useful in loading images from another website via JavaScript, etc.

After some research, it seems that CORS is primarily designed to prevent
in-browser session hijacking.  Linksharing is designed for making things
public, and does not use cookies or session, so I can't find a reason
this would be unsafe.  Ingram is still looking into the topic, but as it
is, we already have similarly permissive CORS setups in Gateway-MT and
Auth Service.

Change-Id: Ic9d3def941c75b3823c63464158346b9a3c7d55c