{testsuite/server,pkg/linksharing}: fix NoError tests to require
this means tests will exit early if they fail, so we don't have to
wrap the subsequent checks in a conditional. We also avoid the chance
of a nil pointer crashing the tests.

TestPutObjectTags, TestGetObjectTags, and TestDeleteObjectTags already
use require.NoError(), so this change makes other tests consistent
with these.

Change-Id: I2daca1a7847bca5a4f3aceb7de94a90dc01f5d2b