authservice: Use rate-limiting failure in getAccess
Use the failure rate-limiting to ban request to the getAccess endpoint
from malicious clients.

Without the rate-limiting a malicious client can cause service outage
making requests to the getAccess endpoint with random well-formatted
access key IDs.

Any random well-formatted access key ID will be used to query the DB;
the query isn't cheap enough to support a very high number of queries,
hence the DB will can be overwhelmed and deny the service to benevolent
clients.

However, the rate-limiting has only to start to apply to clients that
makes a requests with an unexisting access key ID and the limiter
doesn't have to take in account successful requests but, at the same
time, not to allow malicious clients to by-pass it using a combination
of request with unexisting and existing access key IDs.

Last but not least, the commit include a feature flag to enable the rate
limiting (disabled by default) to be able to turn it on/off without
having to release a patch in a case that something gets hairy in
production.

Change-Id: Idcc79cf9b3a9463fa4d1e9838ba62929a6e2693c