{Jenkinsfile,testsuite}: test Gateway-MT running under HTTPS
This adds a test which checks HTTP/2 is listening on HTTPS, in
addition to HTTP 1.1. It also switches over all tests except for
Mint to test against an HTTPS gateway.

For now, Mint tests remain testing against an HTTP gateway as
a number of tests are failing due to SSE functionality returning
501 not implemented errors.

Change-Id: I1c17430551ab644a6eda74b6742ad87a65bfc93c