{pkg/minio,cmd/gateway-mt}: set minio globalIsSSL flag as needed
It was observed that when TLS is enabled on Gateway-MT, minio still
thinks the connection isn't secure when certain requests requiring
a secure connection are made, e.g. SSE parameters on PutObject.

With this change in place, whenever TLS is enabled and SSE requests
are made, the server will correctly respond with a not implemented
error instead of an error about the connection not being secure.

Change-Id: I121648c533bf30747236253daeb84af4a4bf22e4