pkg/minio: replace IAMStorjAuthStore with IAMAuthStore
Authorization was previously handled by IAMStorjAuthStore, which was
required to be within Minio due to the IAMStorageAPI not being fully
exported.

In order to move this code out of Minio, we needed an alternative
approach.  This version uses Minio's IAMObjectStore, which consumes an
ObjectLayer, hence IAMAuthStore implements ObjectLayer.

This version of the code still requires a Minio fork, as
auth.Credentials must still track the AccessGrant.  Future version of
this code may attempt to work around this limitation by using other
Credentials fields, or via alternative means.

Change-Id: I786e108fd2c91ae9ca96ffd911828700df6f54ca