pkg/server: fix path routing to support esoteric characters
This is based on the original fix discussed in minio/minio#8967

The fix itself is already included in storj/minio
with commit id #c56c2f5fd, but the important part was in
the server (cmd/gateway-main.go) where storj has a
different version (server/server.go)

Here the relevant part of the fix is applied to the
gorilla mux in Storj code.

Tested with the rclone/rclone backend/s3 integration tests.

Change-Id: I55d710acdf389c0ff6685e18bcf4f037acf95651