pkg/server: create testable Peer class, largely by moving configs
We want to create Peer structs so that Edge is more like the storj/storj
various high-level structs.  I noticed that the Server struct was very
close to accomplishing this, already having the correct Run() and
Close() methods.

The thorn here is Minio, which uses some global variables, and thus
can only be instantiated once, otherwise there's a data race.

This change changes the integration test to use the testable Gateway-MT
HTTP service (server.(*Peer)) instead of compiling Gateway-MT ad-hoc.

Change-Id: I7c868a14b0bc3db1719ffc284da1280673e5d88c