pkg/minio: remove GlobalEndpoints
This change removes linkname exposure of GlobalEndpoints
(minio/cmd.globalEndpoints) because it's now unused.
GlobalNotificationSys (minio/cmd.globalNotificationSys) can be left as a
global and have passed endpoints as nil because all its methods do
nothing when it's zero-valued. We don't care because we don't use it.
MinIO also doesn't initialise it for gateways except for the NAS
gateway.

Change-Id: I7e7651cc08d4f9dfef21e22244918f9f818fa5ca