pkg: register noop ObjectLayer as Minio global
Gateway-MT now uses a custom RegisterAPIRouter which does not require
that the global ObjectLayer be set in Minio anymore.  This uses a
placeholder layer instead, because Minio errors if the ObjectLayer is
nil.

Multi-planet layer shutdown was cleaned up, because it doesn't make
sense to create a project just to close it.  In reality, shutdown
shouldn't need a project reference or access grant.

Change-Id: If832143efe53370f12922bd3b420ace67c49b978