auth/authdb: add monitoring of satellite address during GETs and PUTs
We want to know if we're actually using the distributed nature of the
auth database.  We should already be tagging all monitoring with the
instance / region, so we just need to log the satellite id to get at
this data.  For purposes of simplicity, the entire satellite address is
logged rather than trying to parse the address.

Change-Id: I1b1731c63270f349a66d991092289679cb1ec560