pkg/server: fix nil pointer if authservice is unavailable
this ensures gateway-mt doesn't error with a nil pointer if the auth
service doesn't respond. Additionally, ensure this is logged so we
know about it. This also means we don't need the minio log targets
anymore, as their reason for existing was to log authservice
connection errors when credentials were being checked by minio,
but we are checking credentials in our middleware now.

Change-Id: Ic29392a61e7e650a2c2df101ae5b1542817d7641