pkg/minio: hacky fix for nil panic
in some cases, like if no access key id is found on the
request, the http request may not have credentials
associated to the request context. in that case, the
GetAccess helper method will return nil. we need to
check for that possibility, otherwise the code panics
and the go net/http server "helpfully" recovers from
it leaving some locks still acquired and eventually
leading to deadlock. oof!

Change-Id: I28ee76430e4bafb1006b6021f9af7409256f4d4d