pkg/linksharing:  add zip file listing and download
Due to the way the Go templates are constructed, this changeset shows
all files in the zip at once, rather than allowing the browsing of one
directory at a time.

Downloads are also supported.  If the ZIP file content is encoded with
deflate, and the client supports gzip, we return gzip content encoding.

Change-Id: Ib5264dc78cb5ff0bdced181782f8cf36a7691845