pkg/auth/sqlauth: revert d176707 and fix the Get method for CRDB v21.2.x
This change partially reverts d17670783d2c8afbb46806088d96825718da2643
and fixes the actual underlying problem of the failing test. The reason
we were failing with CRDB v21.2.x is that it now returns the 3D000 error
code instead of 42P01 when the database does not exist (which happens
when we run AOST queries right after migration). Previously, we only
checked 42P01 before we errored or fell back to a query without AOST.

The d176707-fix was incorrect because we have a fallback to a non-AOST
query implemented.

Reference: https://github.com/cockroachdb/cockroach/pull/71440

Change-Id: Ic4b83d40579c983c8f1f5b21f3afd833d7e6057f