Makefile: clean up docker images properly at end of build
Currently we aren't cleaning up any images because the arch name
suffixed to the end of the docker image tag is not considered.
Because of this, we're potentially leaving around redundant images
on the build servers.

This changes the cleanup commands to find images by wildcard, then
delete these images. `docker rmi` has no way of fuzzy matching, so
we need to pass the list from `docker images`.

Force parameter is needed to ensure the "dev" tagged image is also
removed, otherwise docker will complain the images are linked to
another tag and the command will fail.

Change-Id: I6e20dce56dcaab73be8ae444f545af5593b9cbde