pkg/linksharing: add Twitter and OG meta tags
This change adds Twitter and OG meta tags for social media previews. If
linksharing URL points to an image, this image, if it meets the size and
file format criteria, will be displayed, and a generic preview image
will be used otherwise.

This change is a rework of an old PR (#64).

Closes #64

Change-Id: I7439a72903ada364b07beb572b68afad00e33029