pkg/auth/failrate: Remove monkit tag of counter
Remove the tag passed to a monkit counter metric for avoiding to
overwhelm the service due to a metric that can have too much
cardinality.

The counter is kept but removing the tag, which offers a different and
less useful information (i.e. we won't know which IP gets banned) but
at least we could correlate the counter with the number of entries in
the rate-limiting cache.

Change-Id: I6eaf0125a4cdde3e3f87de0ea4e3daee99d7d725