gateway-mt: refactor to use httpserver package
this changes gateway-mt to use the httpserver package currently
only used by linksharing. this means we don't need to duplicate a
bunch of code, especially around TLS config, and we have the same
start and shutdown logic for both services.

like linksharing, gateway-mt will now listen on both HTTP and
HTTPS, so this change adds the `--server.address-tls` flag, which
defaults to :7778. default ports will be revisited in a later change.

this also adjusts the httpserver run/shutdown logic so that `Run()`
*only* runs the server, and isn't responsible for also shutting it
down. the caller (e.g. a peer) is responsible for shutting down the
server by calling `Shutdown()` such as on signal, or at the end of
a test run.

Updates #118

Change-Id: I316bfa87a40e5aa6ab8c7d21746592e59ba3ef86