pkg/server:  add concurrency limiting for uploads and downloads
Uploads and downloads in Gateway-ST carry a significant memory pressure.
We must limit the "blast radius" of misbehaving clients which create too
many requests.  Uploads and downloads are the only endpoint which use
these large amounts of memory, so the limiter is wired up within the
path-aware API router logic.

One challenge is that Minio's default errors can't apply here. We
compose our own HTTP 429 which I've confirmed show up in the AWS CLI.

Another challenge is determining an appropriate value for the production
limit on how many concurrent requests to allow.  The S3 CLI has a
max_concurrent_requests default of 10, so 500 seems a healthy upper
limit, so as to not produce surprises for existing clients.

Finally, testing showed it was best to not handle errors in the limiter,
but to let Minio handle them (thanks Ceph/Splunk tests!), so that
feature of the limiter has been removed.

Closes storj/gateway-mt#131

Change-Id: If20d6e201e8ae2bf4aa6ceb73dff43d1887fc3dd