cmd/authservice: add "register" command
New uplink CLI got rid of HTTP-based access registration and
additionally changed it to register only securely. Generally, this is
great, but it broke our CI for gateway-mint and splunk-s3-tests
repositories. It seemed like a good idea to make it possible to register
access at authservice with a small, additional command to authservice
executable. This command is not intended for widespread usage (we may
break it, and it's better if people use uplink CLI), so it's hidden from
help. It supports both HTTP and DRPC-based registration depending on the
passed URL's scheme.

Updates #147

Change-Id: I15c7b821b41774e1c8e60826126ab6f88a2359dc