pkg/auth/badgerauth: implement Ping
Implemented ping as reading the start time of the node.
The start time probably isn't useful, but it's the best
I was able to come up with.

Maybe it makes sense to have a separate "Ping" field,
that is being updated on every ping. But, I'm not sure
how often the ping will be called and we wouldn't want to
add undue stress to the database itself.

Change-Id: Ibd8ce32f403c1aa993b35a7a801cca4db3a2b2e0