Jenkinsfile: improve docker cleanup
this ensures all containers within the testing network are cleaned
up at the end of the build, so that we can remove the network without
the "has active endpoints" error. This could happen if any test
containers got stuck when a build timed out.

additionally, use --time=1 to quickly kill unresponsive containers.

Closes storj/gateway-mt#192

Change-Id: If87d8f148913c3cd28ed37be52e0b51991ffb058