pkg/auth/badgerauth: add a chore to garbage collect the value log
This change adds a chore to garbage collect the value log because we
haven't been doing this. In the case of BadgerDB, automatic compactions
are not enough because values are kept separately from the LSM tree.
Value log GC is meant to be run by the user.

Reference: https://dgraph.io/docs/badger/get-started/#garbage-collection

Change-Id: I2df34eed8098e9095311a80d33763d2b75e28ef6