pkg/auth/badgerauth: observe LSM and value log's size with monkit
The change also denotes that we can report more from the underlying
storage engine, but for now, it seems like it will be sufficient to make
decisions based on anomalies in LSM/value log sizes. Further usage of
badgerauth will allow us to curate what more we need to report.

Closes #199

Change-Id: Iaf3a66b2f0931714b5df3b01c467c9a00e94d77f