pkg/server/middleware: provide better error for malformed credentials
If there's an error parsing in our credential middleware, the error
could be important, but it's currently silently discarded. It was
assumed that minio would also see the invalid credential and
respond with an appropriate S3 API error. However, if our parser
failed to parse but minio was able to (e.g. when we failed to allow
an empty region), then minio goes to IAMAuthStore to retrieve an
empty context value, and responds with an access key not found
error which is very misleading.

This changes the middleware to write an S3 error response if it
encountered a problem. It only returns a subset of S3 credential
error responses as part of our "lightweight" approach to parsing
for the access key ID.

Closes storj/gateway-mt#198

Change-Id: I4740fd0c24bef7a6c8cb34451a0fc9a0bf7949cd