pkg/auth/badgerauth: add Peek to ReplicationService
We recently learned that most likely, we'll need to contact other nodes
if the queried node doesn't have searched-for credentials. To achieve
this, we need to make nodes be able to look up specific records at their
peers. With this context in mind, this change adds the Peek method to
ReplicationService, ingesting encryption key hash and returning the
record.

Updates #225

Change-Id: I34588e973d016e619d548bdfb161a692973150de