pkg/auth/badgerauth: broadcast Get
This change broadcasts Get to get records from other nodes concurrently.
Principally in this change, we implement non-guaranteed read after write
consistency for node.

This is a rewrite of
https://review.dev.storj.io/c/storj/gateway-mt/+/8161/13 that uses less
primitive types to make early cancelation a bit easier to understand. It
also waits for all goroutines to finish.

Updates storj/gateway-mt#225

Change-Id: I44acda84d1b3fa493d6015ee0cb111500e1aa794