pkg/auth: change peer to be properly context cancelation aware
This fixes auth peer getting stuck if one of the servers has a
failure. Most likely this would happen if you started authservice
with an address that was already taken. In that case the peer
doesn't error and exit, instead it just stays running in a half
broken state.

This moves the listener creation into New(), so any problems
with those are going to error and exit before we start serving
anything. This is also how we set up listeners in other servers
like storj.io/gateway-mt/pkg/httpserver. Doing it this way means
we don't need to coordinate them starting before calling
SetStartupDone(). This also means we can inspect the addresses
of these listeners once we have a running peer, useful for
finding the random port that was assigned when setting an address
like "127.0.0.1:0"!

This also makes HTTP servers context cancelation aware to match
drpcserver.Server. This means (*Peer).Close() now only closes
the KV backend instead of shutting down HTTP servers, and in order
to shut the peer servers down, you need to cancel the context.
This also means we don't need the stopDRPCAuth cancel stored on
the struct which is a bit odd. Other peers like gateway-mt, and
linksharing probably need to be adjusted to work the same as
these changes, but that can be done later.

Thanks to JT for doing most of the investigation work for this
which helped greatly!

Updates #246

Change-Id: Id33ff1341ee7e53dcd44d170c9d40385bbbe0899