pkg/auth: insert into both stores with equal time during migration
During migration, the following situation can happen:

1. Node A stores record R that gets written to both badgerauth & sqlauth
2. Node B migrates R from sqlauth
3. Node B replicates R from Node A

We can't really avoid this duplication, but it also doesn't hurt.
However, it hurts us when R is inserted with different creation times to
badgerauth and sqlauth. We perform checks in case the record differs
while migrating and replicating, so when that happens, we won't allow
both to proceed.

This change fixes this problem by ensuring we insert to both backends
with the same time truncated to seconds, which is the greatest precision
badgerauth cares about.

Change-Id: Ib4031983db8513670ec25121f9d197cd84d63dd0