pkg/auth/badgerauth: unify start/stop order of Node's services
This change unifies the start/stop order of services that Node must
start to function. This isn't a functional change (except for reordering
of when backups start—they now start right after we open the handle to
the storage engine) because we mostly rely on context's cancelation to
close running services (things that Run start will be closed once passed
context is canceled). However, if there's a chance cancelation didn't
work correctly, or if we start adding more things to Node, it will be
beneficial to have start/stop well-structured.

Change-Id: I023b9419a056219aa0a166451e58f74d980bb776