pkg/linksharing/sharing: prefix URLs in listing with ./
Change-Id: I1403df11d2d0bb34d4f2e9556e27912c53e89d89

Adding ./ prefix to the filepath as a path segment that contains a colon
character cannot be used as the first segment of a relative-path reference, because
it would be mistaken for a scheme name.  Such a segment can be preceded by a dot-segment
`./` to make a relative-path reference.
Reference: https://www.rfc-editor.org/rfc/rfc3986#section-4.2

Change-Id: I388bac511115791edaf77f62b26fd2699c964f18