pkg/auth/badgerauth: make it possible to skip loading certificates
This change makes it possible to skip loading certificates while
initializing the node. It's safe to do because if no certificates are
supplied, empty pools will be returned, and the node will always be
alone in its cluster. In exchange, it will be possible to use badgerauth
replacing memauth for non-prod purposes.

Change-Id: I6e9e3d3016a34691149ec788e537a7490948d3d8