{cmd/authservice,pkg/auth,testsuite/server}: add capability to have multiple authservice tokens
Currently the authservice only allows a single auth token for accessing the sensitive dataset. This creates challenges when different deployments/environments and usecase want to access it or generally we would like to rotate the secret.
This change extends the authservice to allow multiple tokens to be used concurrently, making all the above cases easy to handle.

Change-Id: I0edbec08e4d6c7a4c74f5ad9ab63174287771726