pkg/linksharing: show correct error message if access is invalid
authservice returns HTTP 401 if an access key is missing from the
database, or is invalid. This fixes the Linksharing handler to
check for that status and return the right message instead of
"Internal server error...".

this also fixes tests to correctly mock what the real authservice
is doing: 401 is returned if a key isn't in the db or is invalid.

Updates #305

Change-Id: If7f59b57c3d9b9bdf1733c65fb325834a71fdd08