{pkg/auth,pkg/httpserver,pkg/server/middleware}: adjust log level rules
this change applies the same log rules from gateway-mt to linksharing
and authservice. In that access and response are debug level by
default, info for >= 400 status, and errors for >= 500 status.

this also fixes a bug in that 5xx errors are being logged as info level
incorrectly.

note that successful request events are still sent to eventkit for
linksharing and gateway-mt, so we still have observability over those.

Change-Id: I95f8842d0258acc6f2077774c832fa6aeeadd0b1